+++ script/vm/vm.lua
@@ -21,6 +21,7 @@ function m.getSpecial(source)
     return source.special
 end
 
+---@param source parser.object
 ---@return string?
 function m.getKeyName(source)
     if not source then